app/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
python-multipart
pydantic
PyJWT
passlib[bcrypt]
bcrypt